fix typos, regenerated documentation

+++ packages/acs-core-docs/www/install-openacs-delete-tablespace.adp
@@ -3,11 +3,11 @@
 <property name="doc(title)">Deleting a tablespace</property>
 <master>
 <include src="/packages/acs-core-docs/lib/navheader"
-		    leftLink="remote-postgres" leftLabel="Prev"
-		    title="
-Chapter 7. Database Management"
-		    rightLink="install-next-nightly-vacuum" rightLabel="Next">
-		<div class="sect1">
+			leftLink="remote-postgres" leftLabel="Prev"
+			title="Chapter 7. Database
+Management"
+			rightLink="install-next-nightly-vacuum" rightLabel="Next">
+		    <div class="sect1">
 <div class="titlepage"><div><div><h2 class="title" style="clear: both">
 <a name="install-openacs-delete-tablespace" id="install-openacs-delete-tablespace"></a>Deleting a tablespace</h2></div></div></div><p>Skip down for instructions on <a class="xref" href="install-openacs-delete-tablespace" title="Deleting a PostgreSQL tablespace">Deleting a PostgreSQL
 tablespace</a>.</p><div class="sect2">
@@ -16,20 +16,16 @@ tablespace</a>.</p><div class="sect2">
 tablespace</h3></div></div></div><p>Should it become necessary to rebuild a tablespace from scratch,
 you can use the <code class="computeroutput">drop user</code>
 command in SVRMGRL with the <code class="computeroutput">cascade</code> option. This command will drop the
-user and every database object the user owns.</p><pre class="programlisting">
-SVRMGR&gt; <strong class="userinput"><code>drop user <em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em> cascade;</code></strong>
+user and every database object the user owns.</p><pre class="programlisting">SVRMGR&gt; <strong class="userinput"><code>drop user <em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em> cascade;</code></strong>
 </pre><p>If this does not work because svrmgrl "cannot drop a user
 that is currently connected", make sure to kill the AOLserver
-using this user. If it still does not work, do:</p><pre class="programlisting">
-SVRMGR&gt; <strong class="userinput"><code>select username, sid, serial# from v$session where lower(username)='<em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em>';</code></strong>
-</pre><p>and then</p><pre class="programlisting">
-SVRMGR&gt; <strong class="userinput"><code>alter system kill session '<em class="replaceable"><code>sid, serial#</code></em>';</code></strong>
+using this user. If it still does not work, do:</p><pre class="programlisting">SVRMGR&gt; <strong class="userinput"><code>select username, sid, serial# from v$session where lower(username)='<em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em>';</code></strong>
+</pre><p>and then</p><pre class="programlisting">SVRMGR&gt; <strong class="userinput"><code>alter system kill session '<em class="replaceable"><code>sid, serial#</code></em>';</code></strong>
 </pre><p>where <span class="emphasis"><em>sid</em></span> and
 <span class="emphasis"><em>serial#</em></span> are replaced with
 the corresponding values for the open session.</p><p><span class="strong"><strong>Use with
 caution!</strong></span></p><p>If you feel the need to delete <span class="emphasis"><em>everything</em></span> related to the service, you
-can also issue the following:</p><pre class="programlisting">
-SVRMGR&gt; <strong class="userinput"><code>drop tablespace <em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em> including contents cascade constraints;</code></strong>
+can also issue the following:</p><pre class="programlisting">SVRMGR&gt; <strong class="userinput"><code>drop tablespace <em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em> including contents cascade constraints;</code></strong>
 </pre>
 </div><div class="sect2">
 <div class="titlepage"><div><div><h3 class="title">
@@ -44,14 +40,13 @@ comment out your server in <code class="computeroutput">/etc/inittab</code>, rer
 </code>.</p><p>Then, to drop the db, just do:</p><pre class="programlisting">
 [$OPENACS_SERVICE_NAME ~]$ <strong class="userinput"><code>dropdb <em class="replaceable"><code>$OPENACS_SERVICE_NAME</code></em>
 </code></strong>
-DROP DATABASE
-</pre>
+DROP DATABASE</pre>
 </div>
 </div>
 <include src="/packages/acs-core-docs/lib/navfooter"
-		    leftLink="remote-postgres" leftLabel="Prev" leftTitle="Running a PostgreSQL database on
+			leftLink="remote-postgres" leftLabel="Prev" leftTitle="Running a PostgreSQL database on
 another server"
-		    rightLink="install-next-nightly-vacuum" rightLabel="Next" rightTitle="Vacuum Postgres nightly"
-		    homeLink="index" homeLabel="Home" 
-		    upLink="database-management" upLabel="Up"> 
-		
+			rightLink="install-next-nightly-vacuum" rightLabel="Next" rightTitle="Vacuum Postgres nightly"
+			homeLink="index" homeLabel="Home" 
+			upLink="database-management" upLabel="Up"> 
+
